move click to copy effect to new Container component
