+++ src/components/Main.js
@@ -1,8 +1,4 @@
-// Dependencies
-import { useEffect } from "react";
-import ClipboardJS from "clipboard";
-
-// Section Components
+// Components
 import Popular from "../sections/Popular";
 import Punctuation from "../sections/Punctuation";
 import Accent from "../sections/Accent";
@@ -12,33 +8,6 @@ import Arrows from "../sections/Arrows";
 import Fun from "../sections/Fun";
 
 export default function Main() {
-  // Click to copy
-  useEffect(() => {
-    const items = document.querySelectorAll("li.copyable");
-    const clipboard = new ClipboardJS(items);
-
-    clipboard.on("success", (e) => {
-      // add tooltip
-      e.trigger.insertAdjacentHTML(
-        "beforeend",
-        `<span class="tooltip">Copied!</span>`,
-      );
-      //debugger // use when working on tooltip css
-      // hide tooltip after 2 seconds
-      setTimeout(
-        () =>
-          document
-            .querySelectorAll(".tooltip")
-            .forEach((tooltip) => tooltip.remove()),
-        2000,
-      );
-
-      e.clearSelection();
-    });
-
-    clipboard.on("error", (e) => console.error(e));
-  }, []);
-
   return (
     <main>
       <Popular />
